refactor(de): Improve code quality
