+++ DE/etl-pipeline/common/__init__.py
@@ -1,10 +1,10 @@
-from common.cli import iter_dates_to_today, parse_args, parse_iso_date
+from common.cli import iter_dates, parse_args, parse_iso_date
 from common.db import load_demo_slice
 from common.duckdb_utils import load_sql, sql_bool, sql_number
 from common.paths import resolve_input_path
 
 __all__ = [
-    "iter_dates_to_today",
+    "iter_dates",
     "load_demo_slice",
     "load_sql",
     "parse_args",
