+++ DE/etl-pipeline/features/example/__init__.py
@@ -1,0 +1,3 @@
+from features.example.job import run
+
+__all__ = ["run"]
