--- DE/etl-pipeline/features/feature_examples/__init__.py
@@ -1,3 +1,0 @@
-from features.feature_examples.job import run
-
-__all__ = ["run"]
